split up all visit fns
