+++ src/serializeParameter/visitNode/visitIdentifierExpression.ts
@@ -1,0 +1,17 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+import * as PQP from "@microsoft/powerquery-parser";
+import { DefaultSerializeParameter, SerializeParameterState } from "../types";
+import { maybePropagateWriteKind, propagateWriteKind, setWorkspace } from "./visitNodeUtils";
+
+export function visitIdentifierExpression(
+    state: SerializeParameterState,
+    node: PQP.Language.Ast.IdentifierExpression,
+): void {
+    if (maybePropagateWriteKind(state, node, node.maybeInclusiveConstant)) {
+        setWorkspace(state, node.identifier, DefaultSerializeParameter);
+    } else {
+        propagateWriteKind(state, node, node.identifier);
+    }
+}
