+++ src/serializeParameter/visitNode/visitLeaf.ts
@@ -1,0 +1,31 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+import * as PQP from "@microsoft/powerquery-parser";
+import { SerializeParameter, SerializeParameterState, SerializerWriteKind } from "../types";
+import { visitComments } from "./visitComments";
+import { getWorkspace, maybeSetIndentationChange } from "./visitNodeUtils";
+
+export function visitLeaf(
+    state: SerializeParameterState,
+    node:
+        | PQP.Language.Ast.TConstant
+        | PQP.Language.Ast.GeneralizedIdentifier
+        | PQP.Language.Ast.Identifier
+        | PQP.Language.Ast.LiteralExpression,
+): void {
+    const workspace: SerializeParameter = getWorkspace(state, node);
+    maybeSetIndentationChange(state, node, workspace.maybeIndentationChange);
+
+    let maybeWriteKind: SerializerWriteKind | undefined = workspace.maybeWriteKind;
+    maybeWriteKind = visitComments(state, node, maybeWriteKind);
+    if (!maybeWriteKind) {
+        const details: {} = {
+            node,
+            maybeWriteKind,
+        };
+        throw new PQP.CommonError.InvariantError("maybeWriteKind should be truthy", details);
+    }
+
+    state.result.writeKind.set(node.id, maybeWriteKind);
+}
